reorg scenarios
