+++ test/e2e/framework/azure/delete-rg.go
@@ -27,16 +27,10 @@ func (d *DeleteResourceGroup) Run() error {
 	if err != nil {
 		return fmt.Errorf("failed to create resource group client: %w", err)
 	}
-
 	forceDeleteType := "Microsoft.Compute/virtualMachines,Microsoft.Compute/virtualMachineScaleSets"
-	poller, err := clientFactory.NewResourceGroupsClient().BeginDelete(ctx, d.ResourceGroupName, &armresources.ResourceGroupsClientBeginDeleteOptions{ForceDeletionTypes: to.Ptr(forceDeleteType)})
+	_, err = clientFactory.NewResourceGroupsClient().BeginDelete(ctx, d.ResourceGroupName, &armresources.ResourceGroupsClientBeginDeleteOptions{ForceDeletionTypes: to.Ptr(forceDeleteType)})
 	if err != nil {
 		return fmt.Errorf("failed to finish the delete resource group request: %w", err)
-	}
-
-	_, err = poller.PollUntilDone(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("failed to pull the result for delete resource group: %w", err)
 	}
 
 	log.Printf("resource group \"%s\" deleted successfully", d.ResourceGroupName)
@@ -44,10 +38,6 @@ func (d *DeleteResourceGroup) Run() error {
 }
 
 func (d *DeleteResourceGroup) Prevalidate() error {
-	return nil
-}
-
-func (d *DeleteResourceGroup) Postvalidate() error {
 	return nil
 }
