+++ test/e2e/framework/kubernetes/create-network-policy.go
@@ -72,10 +72,6 @@ func (c *CreateDenyAllNetworkPolicy) Prevalidate() error {
 	return nil
 }
 
-func (c *CreateDenyAllNetworkPolicy) Postvalidate() error {
-	return nil
-}
-
 func (c *CreateDenyAllNetworkPolicy) Stop() error {
 	return nil
 }
@@ -110,9 +106,5 @@ func (d *DeleteDenyAllNetworkPolicy) Run() error {
 }
 
 func (d *DeleteDenyAllNetworkPolicy) Prevalidate() error {
-	return nil
-}
-
-func (d *DeleteDenyAllNetworkPolicy) Postvalidate() error {
 	return nil
 }
