+++ test/e2e/framework/kubernetes/exec-pod.go
@@ -39,10 +39,6 @@ func (e *ExecInPod) Prevalidate() error {
 	return nil
 }
 
-func (e *ExecInPod) Postvalidate() error {
-	return nil
-}
-
 func (e *ExecInPod) Stop() error {
 	return nil
 }
