+++ test/e2e/framework/types/runner.go
@@ -6,6 +6,8 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+// A wrapper around a job, so that internal job components don't require things like *testing.T
+// and can be reused elsewhere
 type Runner struct {
 	t   *testing.T
 	Job *Job
