+++ test/e2e/framework/types/step.go
@@ -1,14 +1,27 @@
 package types
 
 var DefaultOpts = StepOptions{
-	ExpectError:               false,
+	// when wanting to expect an error, set to true
+	ExpectError: false,
+
+	// when wanting to avoid saving the parameters to the job,
+	// such as a repetetive task where step is used multiple times sequentially,
+	// but parameters are different each time
 	SkipSavingParamatersToJob: false,
 }
 
 type Step interface {
+	// Useful when wanting to do parameter checking, for example
+	// if a parameter length is known to be required less than 80 characters,
+	// do this here so we don't find out later on when we run the step
+	// when possible, try to avoid making external calls, this should be fast and simple
 	Prevalidate() error
+
+	// Primary step where test logic is executed
+	// Returning an error will cause the test to fail
 	Run() error
-	Postvalidate() error
+
+	// Require for background steps
 	Stop() error
 }
