+++ test/e2e/framework/types/step_sleep.go
@@ -22,7 +22,3 @@ func (c *Sleep) Stop() error {
 func (c *Sleep) Prevalidate() error {
 	return nil
 }
-
-func (c *Sleep) Postvalidate() error {
-	return nil
-}
