--- test/e2e/scenarios/cns/scenarios.go
@@ -1,3 +1,0 @@
-package cns
-
-// reserved
